// src/app/vacancy/interfaces/SupportStaffByVacancyResponse.interface.ts
export default interface SupportStaffByVacancyResponse {
    id:                    number;
    idunique_vacancy:      string;
    number_openings:       number;
    assignment_date:       string;
    approval_target:       number;
    responsible:           number;
    role_responsible:      number;
    responsible_name:      string;
    role_responsible_name: string;
}